removed closure throw and tracked issue for further investigation

+++ Core/IntegrationTests/ServerTransportTests.cs
@@ -1,3 +1,4 @@
+using System.Collections;
 using Speckle.Core.Api;
 using Speckle.Core.Credentials;
 using Speckle.Core.Helpers;
@@ -45,11 +46,10 @@ public class ServerTransportTests
 
   private void CleanData()
   {
-    try
+    if (Directory.Exists(_basePath))
     {
       Directory.Delete(_basePath, true);
     }
-    catch (DirectoryNotFoundException) { }
   }
 
   [Test]
@@ -141,17 +141,18 @@ public class ServerTransportTests
       .ConfigureAwait(false);
 
     memTransport = new MemoryTransport();
-    var receivedObject = await Operations
+    Base? receivedObject = await Operations
       .Receive(
         sentObjectId,
         transport,
         memTransport,
         onErrorAction: (s, e) =>
         {
-          Console.WriteLine(s);
+          throw new Exception(s, e);
         }
       )
       .ConfigureAwait(false);
+    Assert.That(receivedObject, Is.Not.Null);
 
     var allFiles = Directory
       .GetFiles(transport.BlobStorageFolder)
@@ -164,7 +165,7 @@ public class ServerTransportTests
     // Check that there are three downloaded blobs!
     Assert.That(blobPaths.Count, Is.EqualTo(3));
 
-    var blobs = (receivedObject["blobs"] as IList<object>).Cast<Blob>().ToList();
+    var blobs = ((IList)receivedObject!["blobs"]!).Cast<Blob>().ToList();
     // Check that we have three blobs
     Assert.IsTrue(blobs.Count == 3);
     // Check that received blobs point to local path (where they were received)
